move components to own directories
