+++ website/src/routes/_benchmarks/json-schema/styles.css
@@ -1,5 +1,5 @@
-@import "./conversion/-components/table.css";
-@import "./conversion/-components/card.css";
+@import "./conversion/-components/table/index.css";
+@import "./conversion/-components/card/index.css";
 
 .json-schema-cards {
   display: grid;
